# Increase test log level to 'warn'