This is a shameless tweak to improve performance when running tests
in Docker, which is sensitive to heavy disk IO. I'm pretty sure we
don't have a use case for info-level logs in tests. I'm also open to
making this an environment variable, which might be a more reusable
approach across some of the other apps. I'm seeing about a 30-40s
improvement in test runtime on my machine.

Before

    govuk-docker-run bundle exec rspec  1.59s user 0.27s system 0% cpu 3:44.85 total

After

    govuk-docker-run bundle exec rspec  1.55s user 0.27s system 0% cpu 3:02.77 total