@kevindew 2020-02-05
⭐️ for providing benchmarks.

That's a great improvement 👍 

I don't have any qualms with doing this as I think it's incredibly low the amount of times I benefit from info level logs in tests. I don't think we need an env var either really as I imagine in the rare scenario you'd want to change this you could just edit the file - I imagine mostly people reach for a debugger before considering that.

On my local machine I see a much smaller drop in time but still a drop:

Before: Finished in 1 minute 59 seconds (files took 6.12 seconds to load)
After: Finished in 1 minute 52.36 seconds (files took 5.73 seconds to load)

Have you experimented with making the log directory delegated on docker to see if that helps performance at all? Also, although not a practical long term thing, have you tried mounting the logs and/or tmp on a tempfs to see how big an impact that has?