@kevindew 2020-02-05
> On my machine I did a cold run on Ben's branch, finishing in 3m47s. Then switched to master, where it completed in 3m01s 😕
> 
> But spoke to Ben and we figured the second run must have benefited from some caching. When I re-ran the tests in Ben's branch, it came down to 2m26.5s 🎉

Yup that time difference looks likely to be the time it takes to build the assets 🤢