@benthorner 2020-02-05
Thanks @kevindew and @ChrisBAshton for trying this out. Here's some more benchmarks for my machine.

After (again)

    govuk-docker-run bundle exec rspec  1.56s user 0.27s system 1% cpu 2:54.57 total

Before (again)

    govuk-docker-run bundle exec rspec  1.65s user 0.38s system 0% cpu 3:52.63 total

After (+tmpfs stuff)

    govuk-docker-run bundle exec rspec  1.63s user 0.28s system 1% cpu 2:58.19 total

@kevindew I'd been toying with tmpfs stuff (it isn't clear from the docs that it works with Macs). For logs, it would be a shame to make them completely unavailable. FYI you need to explicitly set the `uid` of the mount point to `1000` in order for it to work properly i.e.

```
tmpfs:
  - /govuk/content-publisher/tmp/storage:uid=1000
  - /govuk/content-publisher/log:uid=1000
```

It looks like it doesn't provide any particular improvements over the log approach. And the Redis issue is new to me 🤔.